Add collection products feature with navigation

+++ Buyza App/Buyza App/Presentation/Home/ViewModel/HomeViewModel.swift
@@ -70,9 +70,9 @@ class HomeViewModel : ObservableObject {
         }
     }
     
-    func isFavorite(productID: Int) -> Bool {
+    func isFavorite(productID: Int64) -> Bool {
         // will call here the function from the usecase
-        
+
         return true
     }
 }
